Fixed race condition
There was a race where several status updates entered
the stream before our controller had a chance to process
and annotate it. The fix is to add a `resourceVersion`
field when we patch the object so that the patch API
call will fail if there's a newer version of the object
in the system. We will eventually process the newer object

Also, replaced the objects we perform on from Job to Pod
to simplify the example and added an ASCII art for
coffee_handler.